tradeiq_backend/requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.0.0
gunicorn==20.1.0
whitenoise==6.5.0
yfinance==1.2.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.0
requests==2.32.3
reportlab==4.2.5
psycopg2-binary==2.9.10
bcrypt==4.3.0
setuptools<81
